flip from number to other

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -10,7 +10,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 int i = 0;
-unsigned int a;
+unsigned long int a;
 
 a = n ^ m;
